backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
supabase==2.11.0

# Machine Learning
xgboost==2.1.3
scikit-learn==1.6.1
shap==0.46.0

# Experiment Tracking
mlflow==2.19.0

# Data Processing
pandas==2.2.3
numpy==2.2.1

# Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3
selenium==4.27.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
joblib==1.4.2
